requirements.txt
streamlit
google-generativeai
pypdf
langchain
langchain-google-genai
faiss-cpu